Move field definition prioritization to utils
