+++ MedLog/frontend/components/Drug/ResultCard.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { computed, ref } from "#imports";
 import type {
-  SchemaDisplayPriorityClass,
   SchemaDrugAttrFieldDefinitionContainer,
   SchemaDrugCodeSystem,
   SchemaMedLogSearchEngineResult,
@@ -74,36 +73,7 @@ const keyValuePills = computed(() =>  {
 });
 
 const prioritizedFieldDefinitions = computed(() => {
-  const fields: Record<SchemaDisplayPriorityClass, { attributeClass: keyof SchemaDrugAttrFieldDefinitionContainer, fieldDefinition: FieldDefinition }[]> = { 1: [], 2: [], 3: [] };
-
-  // Group field definitions by priority class
-  for (const attributeClassString of Object.keys(drugFieldsStore.fieldsForSearchResults)) {
-    const attributeClass = attributeClassString as keyof typeof drugFieldsStore.fieldsForSearchResults;
-    for (const fieldDefinition of drugFieldsStore.fieldsForSearchResults[attributeClass] as FieldDefinition[]) {
-      if (fieldDefinition.field_display_priority_class && Object.keys(fields).includes(String(fieldDefinition.field_display_priority_class))) {
-        fields[fieldDefinition.field_display_priority_class].push({ attributeClass, fieldDefinition });
-      }
-    }
-  }
-
-  // Sort within priority classes
-  for (const priorityClassArray of Object.values(fields)) {
-    priorityClassArray.sort((a, b) => {
-      if (a.fieldDefinition.field_display_sort_order == b.fieldDefinition.field_display_sort_order) {
-        return 0;
-      }
-
-      if (a.fieldDefinition.field_display_sort_order === null) {
-        return 1;
-      } else if (b.fieldDefinition.field_display_sort_order === null) {
-        return -1;
-      }
-
-      return a.fieldDefinition.field_display_sort_order - b.fieldDefinition.field_display_sort_order;
-    });
-  }
-
-  return fields;
+  return useGetPrioritizedFieldDefinitions(drugFieldsStore.fieldsForSearchResults);
 });
 
 const accessDate = computed(() => Date.parse(props.drug.market_access_date ?? ""));
